records: merge disambiguation receivers
